# backend/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Whitfield WMS API",
  "version": "1.0.0"
 },
 "paths": {
...
  "/v1/audit-logs": {
   "get": {
    "tags": [
     "Audit Logs"
    ],
    "summary": "List Audit Logs",
    "description": "List audit log entries (admin only).\n\nArgs:\n    record_id (Optional[str]): Record id filter.\n    limit (int): Max entries.\n    auth (dict): Authenticated user claims.\n\nReturns:\n    list[AuditLogResponse]: Audit entries.",
    "operationId": "list_audit_logs_v1_audit_logs_get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "record_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Record Id"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 500,
       "minimum": 1,
       "default": 50,
       "title": "Limit"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/AuditLogResponse"
         },
         "title": "Response List Audit Logs V1 Audit Logs Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/audit-logs/history": {
   "get": {
    "tags": [
     "Audit Logs"
    ],
    "summary": "Audit History",
    "description": "Return audit history for a specific record (admin only).\n\nArgs:\n    record_id (str): Record id.\n    collection_name (str): Collection name.\n    auth (dict): Authenticated user claims.\n\nReturns:\n    list[AuditLogResponse]: Audit history.",
    "operationId": "audit_history_v1_audit_logs_history_get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "record_id",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Record Id"
      }
     },
     {
      "name": "collection_name",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Collection Name"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/AuditLogResponse"
         },
         "title": "Response Audit History V1 Audit Logs History Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/damage-records": {
   "post": {
    "tags": [
     "Damage Records"
    ],
    "summary": "Create Damage Record",
    "description": "Create a damage assessment record.\n\nArgs:\n    request (DamageRecordCreate): Damage record payload.\n    auth (dict): Authenticated user claims.\n\nReturns:\n    DamageRecordResponse: Created damage record.",
    "operationId": "create_damage_record_v1_damage_records_post",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/DamageRecordCreate"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/DamageRecordResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "Damage Records"
    ],
    "summary": "List Damage Records",
    "description": "List damage records.\n\nArgs:\n    warehouse_id (Optional[str]): Warehouse filter.\n    auth (dict): Authenticated user claims.\n\nReturns:\n    list[DamageRecordResponse]: Damage records.",
    "operationId": "list_damage_records_v1_damage_records_get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "warehouse_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Warehouse Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/DamageRecordResponse"
         },
         "title": "Response List Damage Records V1 Damage Records Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/damage-records/{damage_id}": {
   "get": {
    "tags": [
     "Damage Records"
    ],
    "summary": "Get Damage Record",
    "description": "Fetch a damage record by id.\n\nArgs:\n    damage_id (str): Damage record id.\n    auth (dict): Authenticated user claims.\n\nReturns:\n    DamageRecordResponse: Damage record.",
    "operationId": "get_damage_record_v1_damage_records__damage_id__get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "damage_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Damage Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/DamageRecordResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/bin-locations": {
   "post": {
    "tags": [
     "Bin Locations"
    ],
    "summary": "Create Bin Location",
    "description": "Create a bin location.\n\nArgs:\n    request (BinLocationCreate): Bin location payload.\n    auth (dict): Authenticated user claims.\n\nReturns:\n    BinLocationResponse: Created bin location.",
    "operationId": "create_bin_location_v1_bin_locations_post",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/BinLocationCreate"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/BinLocationResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "Bin Locations"
    ],
    "summary": "List Bin Locations",
    "description": "List bin locations.\n\nArgs:\n    warehouse_id (Optional[str]): Warehouse filter.\n    empty (bool): Only empty bins.\n    auth (dict): Authenticated user claims.\n\nReturns:\n    list[BinLocationResponse]: Bin locations.",
    "operationId": "list_bin_locations_v1_bin_locations_get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "warehouse_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Warehouse Id"
      }
     },
     {
      "name": "empty",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": false,
       "title": "Empty"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/BinLocationResponse"
         },
         "title": "Response List Bin Locations V1 Bin Locations Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/bin-locations/product/{product_id}": {
   "get": {
    "tags": [
     "Bin Locations"
    ],
    "summary": "Find Bin For Product",
    "description": "Find the bin where a product is stored.\n\nArgs:\n    product_id (str): Product id.\n    warehouse_id (str): Warehouse id.\n    auth (dict): Authenticated user claims.\n\nReturns:\n    BinLocationResponse: Bin location.",
    "operationId": "find_bin_for_product_v1_bin_locations_product__product_id__get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "product_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Product Id"
      }
     },
     {
      "name": "warehouse_id",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Warehouse Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/BinLocationResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/bin-locations/{bin_id}": {
   "patch": {
    "tags": [
     "Bin Locations"
    ],
    "summary": "Update Bin Location",
    "description": "Update a bin location.\n\nArgs:\n    bin_id (str): Bin location id.\n    request (BinLocationUpdate): Update payload.\n    auth (dict): Authenticated user claims.\n\nReturns:\n    BinLocationResponse: Updated bin location.",
    "operationId": "update_bin_location_v1_bin_locations__bin_id__patch",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "bin_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Bin Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/BinLocationUpdate"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/BinLocationResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
   "AuditLogResponse": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "user_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "User Id"
     },
     "user_name": {
      "type": "string",
      "title": "User Name"
     },
     "action": {
      "type": "string",
      "title": "Action"
     },
     "collection_name": {
      "type": "string",
      "title": "Collection Name"
     },
     "record_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Record Id"
     },
     "warehouse_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Warehouse Id"
     },
     "old_value": {
      "type": "object",
      "title": "Old Value"
     },
     "new_value": {
      "type": "object",
      "title": "New Value"
     },
     "method": {
      "$ref": "#/components/schemas/AuditMethod"
     },
     "ip_address": {
      "type": "string",
      "title": "Ip Address"
     },
     "created_at": {
      "type": "string",
      "title": "Created At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "user_name",
     "action",
     "collection_name",
     "old_value",
     "new_value",
     "method",
     "ip_address",
     "created_at"
    ],
    "title": "AuditLogResponse",
    "description": "Audit log entry representation."
   },
   "AuditMethod": {
    "type": "string",
    "enum": [
     "manual_entry",
     "barcode_scan",
     "voice_input",
     "ai_agent",
     "api_script",
     "system_auto"
    ],
    "title": "AuditMethod",
    "description": "Source method of a recorded action."
   },
...
   "BinLocationCreate": {
    "properties": {
     "warehouse_id": {
      "type": "string",
      "title": "Warehouse Id"
     },
     "bin_code": {
      "type": "string",
      "title": "Bin Code"
     },
     "aisle": {
      "type": "string",
      "title": "Aisle"
     },
     "row": {
      "type": "string",
      "title": "Row"
     },
     "shelf": {
      "type": "string",
      "title": "Shelf"
     },
     "bin": {
      "type": "string",
      "title": "Bin"
     },
     "product_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Product Id"
     },
     "max_capacity": {
      "type": "integer",
      "title": "Max Capacity",
      "default": 100
     },
     "current_units": {
      "type": "integer",
      "title": "Current Units",
      "default": 0
     },
     "is_occupied": {
      "type": "boolean",
      "title": "Is Occupied",
      "default": false
     }
    },
    "type": "object",
    "required": [
     "warehouse_id",
     "bin_code",
     "aisle",
     "row",
     "shelf",
     "bin"
    ],
    "title": "BinLocationCreate",
    "description": "Payload to create a bin location."
   },
   "BinLocationResponse": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "warehouse_id": {
      "type": "string",
      "title": "Warehouse Id"
     },
     "bin_code": {
      "type": "string",
      "title": "Bin Code"
     },
     "aisle": {
      "type": "string",
      "title": "Aisle"
     },
     "row": {
      "type": "string",
      "title": "Row"
     },
     "shelf": {
      "type": "string",
      "title": "Shelf"
     },
     "bin": {
      "type": "string",
      "title": "Bin"
     },
     "product_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Product Id"
     },
     "max_capacity": {
      "type": "integer",
      "title": "Max Capacity"
     },
     "current_units": {
      "type": "integer",
      "title": "Current Units"
     },
     "is_occupied": {
      "type": "boolean",
      "title": "Is Occupied"
     }
    },
    "type": "object",
    "required": [
     "id",
     "warehouse_id",
     "bin_code",
     "aisle",
     "row",
     "shelf",
     "bin",
     "max_capacity",
     "current_units",
     "is_occupied"
    ],
    "title": "BinLocationResponse",
    "description": "Bin location representation."
   },
   "BinLocationUpdate": {
    "properties": {
     "product_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Product Id"
     },
     "max_capacity": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Max Capacity"
     },
     "current_units": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Current Units"
     },
     "is_occupied": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Is Occupied"
     }
    },
    "type": "object",
    "title": "BinLocationUpdate",
    "description": "Payload to update a bin location."
   },
...
   "DamageGrade": {
    "type": "string",
    "enum": [
     "A",
     "B",
     "C",
     "D"
    ],
    "title": "DamageGrade",
    "description": "Standardized four-level damage grading system."
   },
   "DamageRecordCreate": {
    "properties": {
     "shipment_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Shipment Id"
     },
     "shipment_ref": {
      "type": "string",
      "title": "Shipment Ref",
      "default": ""
     },
     "product_id": {
      "type": "string",
      "title": "Product Id"
     },
     "warehouse_id": {
      "type": "string",
      "title": "Warehouse Id"
     },
     "quantity_damaged": {
      "type": "integer",
      "minimum": 1.0,
      "title": "Quantity Damaged"
     },
     "damage_grade": {
      "$ref": "#/components/schemas/DamageGrade"
     },
     "damage_notes": {
      "type": "string",
      "title": "Damage Notes",
      "default": ""
     },
     "carrier": {
      "type": "string",
      "title": "Carrier",
      "default": ""
     },
     "carrier_tracking": {
      "type": "string",
      "title": "Carrier Tracking",
      "default": ""
     },
     "action_taken": {
      "type": "string",
      "title": "Action Taken",
      "default": ""
     }
    },
    "type": "object",
    "required": [
     "product_id",
     "warehouse_id",
     "quantity_damaged",
     "damage_grade"
    ],
    "title": "DamageRecordCreate",
    "description": "Payload to create a damage assessment."
   },
   "DamageRecordResponse": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "shipment_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Shipment Id"
     },
     "shipment_ref": {
      "type": "string",
      "title": "Shipment Ref"
     },
     "product_id": {
      "type": "string",
      "title": "Product Id"
     },
     "product_name": {
      "type": "string",
      "title": "Product Name"
     },
     "seller_id": {
      "type": "string",
      "title": "Seller Id"
     },
     "warehouse_id": {
      "type": "string",
      "title": "Warehouse Id"
     },
     "quantity_damaged": {
      "type": "integer",
      "title": "Quantity Damaged"
     },
     "damage_grade": {
      "$ref": "#/components/schemas/DamageGrade"
     },
     "damage_notes": {
      "type": "string",
      "title": "Damage Notes"
     },
     "carrier": {
      "type": "string",
      "title": "Carrier"
     },
     "carrier_tracking": {
      "type": "string",
      "title": "Carrier Tracking"
     },
     "assessed_by": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Assessed By"
     },
     "action_taken": {
      "type": "string",
      "title": "Action Taken"
     },
     "seller_notified": {
      "type": "boolean",
      "title": "Seller Notified"
     }
    },
    "type": "object",
    "required": [
     "id",
     "shipment_ref",
     "product_id",
     "product_name",
     "seller_id",
     "warehouse_id",
     "quantity_damaged",
     "damage_grade",
     "damage_notes",
     "carrier",
     "carrier_tracking",
     "action_taken",
     "seller_notified"
    ],
    "title": "DamageRecordResponse",
    "description": "Damage assessment representation."
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  },
  "securitySchemes": {
   "OAuth2PasswordBearer": {
    "type": "oauth2",
    "flows": {
     "password": {
      "scopes": {},
      "tokenUrl": "/v1/auth/login"
     }
    }
   }
  }
 }
}